第一次提交

+++ ecommerce-order-service/src/main/java/example/ecommerce;/entity/MainOrder.java
@@ -32,7 +32,7 @@ public class MainOrder implements Serializable {
 
     private Long userId;
 
-    private BigDecimal totalAmount;
+    private Integer totalAmount;
 
     private Byte status;
